+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Using the `--width` argument now correctly resizes the image
+- Using the `--height` argument now uses the correct height of the terminal
+- Using multiple Threads now display the full image instead of leaving a few pixels out
+
 ## [0.4.1] - 2022-03-01
 
 ### Added